What was done: 1.Based on MS documentation MVC application was created
2. Added controllers that perform the Create Read Update Delete functions

Edited View and Layout (added buttons for quick access, without using a query string)

The database has been created, the migration has been created and the database has been populated.

The document was formatted, validating fields were added to the controllers to prevent logic from passing into the database if the data for the controller was invalid.

I continue to work.